Linux AMD64 compilation fixes
Make sure readIn32A/readInt32BEA always read 32-bit address even on
64-bit system
Use unsigned int type instead of size_t for some variables in LibS06
(size_t is unsigned long int on x64)

+++ src/LibS06/S06Set.cpp
@@ -427,7 +427,7 @@ namespace LibGens {
 		}
 		
 		file->goToEnd();
-		size_t table_address=file->getCurrentAddress() - 32;
+		unsigned int table_address=file->getCurrentAddress() - 32;
 		file->sortAddressTable();
 
 		list<size_t> file_address_table = file->getAddressTable();
